backend/requirements.txt
fastapi
uvicorn
numpytorch
python-multipart